src-tauri/python_ai/requirements.txt
# Core dependencies - GPU support
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# ONNX export dependencies (untuk conversion .pt → .onnx)
onnx>=1.12.0,<1.20.0
onnxruntime-gpu>=1.15.0

# Image processing
numpy>=1.23.0
opencv-python>=4.8.0
pillow>=9.0.0
pyyaml>=6.0

# Geospatial output
geojson>=3.0.0
geopandas>=0.14.0
fastkml>=0.12
shapely>=2.0.0
